requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pandas==2.1.3
openpyxl==3.1.2
duckduckgo-search==4.1.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.24.3
scikit-learn==1.3.2
openai==1.3.7
